requirements.txt
dicttoxml==1.3.1
Django==1.8.2
django-crispy-forms==1.4.0
django-extensions==1.4.9
django-filter==0.10.0
django-filters==0.1.0
django-floppyforms==1.4.1
django-geojson==2.5.0
django-import-export==0.2.7
django-jsonview==0.4.3
django-leaflet==0.13.4
django-localflavor==1.0
django-mathfilters==0.3.0
django-multiupload==0.3
django-parsley==0.5
django-registration==1.0
django-report-builder==3.1.8
django-report-utils==0.3.1
django-rest-swagger==0.2.6
django-tables2==1.0.4
djangorestframework==3.1.3
gdata==2.0.18
google-api-python-client==1.2
httplib2==0.9
jdcal==1.0
macholib==1.5.1
modulegraph==0.10.4
mongoengine==0.8.7
MySQL-python==1.2.5
nose==1.3.4
oauth2client==1.2
openpyxl==2.1.5
Pillow==2.5.1
py2app==0.7.3
Pygments==2.0.1
pyOpenSSL==0.13.1
pyparsing==2.0.1
python-dateutil==1.5
pytz==2013.7
PyYAML==3.11
six==1.9.0
sqlparse==0.1.15
tablib==0.10.0
the==0.1.5
tornado==4.1
Twisted==13.2.0
requests==2.7.0
requests-oauth==0.4.1
requests-oauthlib==0.5.0
python-social-auth==0.2.12
gunicorn==19.3.0
